Move result class from client to output package.
